offers: refactor code into a modular structure
Split lndk_offers.rs into a proper module structure with separate files:
- handler:      Defines entry point to get and pay bolt12 invoices and
                pay offers (does both). Also actions whenever a offer
                related onion message arrives to lndk.
- lnd_requests: Defines parsing and request to the LND node. Previously
                some were were bounded to the lndk_offer. Now they are
                separated and mostly use traits for easy mocking.
- client_impls: Defines client traits to be used on requests. Also adds
                mocks to them exposed for use in other parts of the
		module.
- parse:        General bolt12 offers parsing methods.
- mod:          General module, exposes Errors and different modules.

This separation improves maintainability and code organization while keeping
the same behavior.
