+++ src/lnd.rs
@@ -1,4 +1,4 @@
-use crate::OfferError;
+use crate::offers::OfferError;
 use async_trait::async_trait;
 use bitcoin::hashes::sha256;
 use bitcoin::hashes::Hash;
@@ -401,6 +401,17 @@ pub async fn build_seed_from_lnd_node(signer: &mut impl MessageSigner) -> Result
     let hash = sha256::Hash::hash(&signature);
     let seed = hash.to_byte_array();
     Ok(seed)
+}
+
+/// Converts vector of bits numbers as i32 (0 - 128) to little endian bytes
+fn feature_bits_to_le_bytes(feature_bits: &[i32]) -> Vec<u8> {
+    let mut bits: u128 = 0;
+    for &bit in feature_bits {
+        if (0..=127).contains(&bit) {
+            bits |= 1u128 << bit;
+        }
+    }
+    bits.to_le_bytes().to_vec()
 }
 
 /// MessageSigner provides a layer of abstraction over the LND API for message signing.
