+++ src/offers/mod.rs
@@ -1,0 +1,84 @@
+use std::{error::Error, fmt::Display};
+
+use lightning::{
+    ln::channelmanager::PaymentId,
+    offers::{merkle::SignError, parse::Bolt12SemanticError},
+};
+use tonic_lnd::tonic::Status;
+
+mod client_impls;
+pub mod handler;
+mod lnd_requests;
+mod parse;
+
+pub use parse::{decode, get_destination, validate_amount};
+
+#[derive(Debug)]
+/// OfferError is an error that occurs during the process of paying an offer.
+pub enum OfferError {
+    /// AlreadyProcessing indicates that we're already trying to make a payment with the same id.
+    AlreadyProcessing(PaymentId),
+    /// BuildUIRFailure indicates a failure to build the unsigned invoice request.
+    BuildUIRFailure(Bolt12SemanticError),
+    /// SignError indicates a failure to sign the invoice request.
+    SignError(SignError),
+    /// DeriveKeyFailure indicates a failure to derive key for signing the invoice request.
+    DeriveKeyFailure(Status),
+    /// User provided an invalid amount.
+    InvalidAmount(String),
+    /// Invalid currency contained in the offer.
+    InvalidCurrency,
+    /// Unable to connect to peer.
+    PeerConnectError(Status),
+    /// No node address.
+    NodeAddressNotFound,
+    /// Cannot list peers.
+    ListPeersFailure(Status),
+    /// Failure to build a reply path.
+    BuildBlindedPathFailure,
+    /// Unable to find or send to payment route.
+    RouteFailure(Status),
+    /// Failed to track payment.
+    TrackFailure(Status),
+    /// Failed to send payment.
+    PaymentFailure,
+    /// Failed to receive an invoice back from offer creator before the timeout.
+    InvoiceTimeout(u32),
+    /// Failed to find introduction node for blinded path.
+    IntroductionNodeNotFound,
+    /// Cannot fetch channel info.
+    GetChannelInfo(Status),
+}
+
+impl Display for OfferError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            OfferError::AlreadyProcessing(id) => {
+                write!(
+                    f,
+                    "We're already trying to pay for a payment with this id {id}"
+                )
+            }
+            OfferError::BuildUIRFailure(e) => write!(f, "Error building invoice request: {e:?}"),
+            OfferError::SignError(e) => write!(f, "Error signing invoice request: {e:?}"),
+            OfferError::DeriveKeyFailure(e) => write!(f, "Error signing invoice request: {e:?}"),
+            OfferError::InvalidAmount(e) => write!(f, "User provided an invalid amount: {e:?}"),
+            OfferError::InvalidCurrency => write!(
+                f,
+                "LNDK doesn't yet support offer currencies other than bitcoin"
+            ),
+            OfferError::PeerConnectError(e) => write!(f, "Error connecting to peer: {e:?}"),
+            OfferError::NodeAddressNotFound => write!(f, "Couldn't get node address"),
+            OfferError::ListPeersFailure(e) => write!(f, "Error listing peers: {e:?}"),
+            OfferError::BuildBlindedPathFailure => write!(f, "Error building blinded path"),
+            OfferError::RouteFailure(e) => write!(f, "Error routing payment: {e:?}"),
+            OfferError::TrackFailure(e) => write!(f, "Error tracking payment: {e:?}"),
+            OfferError::PaymentFailure => write!(f, "Failed to send payment"),
+            OfferError::InvoiceTimeout(e) => write!(f, "Did not receive invoice in {e:?} seconds."),
+            OfferError::IntroductionNodeNotFound => write!(f, "Could not find introduction node."),
+            OfferError::GetChannelInfo(e) => write!(f, "Could not fetch channel info: {e:?}"),
+        }
+    }
+}
+
+impl Error for OfferError {}
